Use field instead of getter
After examining the context of the changes I made, I still think it makes more sense to me if `findKeeper` takes a class instead of an instance.

Though convoluted to beginners, this section is named "Using Class Types in Generics", which, as far as I understand, means restricting arguments of class type using generic.

Moreover, in the text before the code snippet that I made changes to, it has been explicitly indicated that "A more advanced example uses the prototype property to infer and constrain relationships between the constructor function and the instance side of class types.". 

So I would suggest still taking a class as argument instead of an instance, or removing this part at all if it's considered too convoluted for the level of this handbook, so that the logic of this section can be consistent.