Update file extension from `.js` to `.ts` in "Module Augmentation".
It seems to me that the use of `.js` is maybe intentional on the `map.js` extension, but questionable on the `observable.js` since that example specifies a type variable (i.e. `Observable<T>`), an attribute which would render it invalid JavaScript.

I'm not sure if having the first examples use `.js` extensions is adding much value, but I do think that switching file extensions between blocks might be a subtle nuance that could confuse the reader, so this changes both `.js` extensions to use `.ts`.

Ref: https://www.typescriptlang.org/docs/handbook/declaration-merging.html#module-augmentation
Fixes: https://github.com/Microsoft/TypeScript-Handbook/issues/562 (closed)